moved main_contract_tests to a different file
